Parsing synthesis timing estimation HTML report instead of the full pnr final report Parsing the html report was both easy and annoying at the same time. WTH Gowin?!?! Will the 'old' timing report parsing code have any use at some point? Might just nuke it.

+++ examples/tool_tests/gowin.c
@@ -9,7 +9,7 @@
 #include "uintN_t.h"
 #pragma MAIN my_pipeline
 // @tiagosr: sized ports down to fit the amount of I/O in the device
-uint16_t my_pipeline(uint16_t x, uint16_t y)
+uint32_t my_pipeline(uint32_t x, uint32_t y)
 {
   return x + y;
 }
